Criando controler pedido

+++ src/main/resources/db/migration/V4__create_usuario_endereco_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE usuario_endereco (
-	id_usuario int8 NOT NULL,
-	id_endereco int8 NOT NULL,
+	id_usuario integer NOT NULL,
+	id_endereco integer NOT NULL,
 	CONSTRAINT fk_endereco FOREIGN KEY (id_endereco) REFERENCES endereco(id),
 	CONSTRAINT fk_usuario FOREIGN KEY (id_usuario) REFERENCES usuario(id)
 );
